refactor: replace ContactSidePanelHeader with SidePanelContactHeader and add SidePanelContactInfo component
